pyproject.toml
[project]
name = "adelphos-ap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.125.0",
    "requests>=2.32.5",
]
